Move DumpRing meta to a separate file
It was strange that it was in dataplane/config.h
