+++ dataplane/config.h
@@ -78,31 +78,6 @@ struct tDataPlaneConfig
 		SERIALIZABLE(tag, core_id, socket_id);
 	};
 
-	enum class RingMode : uint8_t
-	{
-		Read,
-		Stop,
-		Follow
-	};
-
-	/**
-	 * 64-byte lock-free header that sits in front of each pcap ring in SHM.
-	 *
-	 * ┌────────────┬────────────┬────────────────┐
-	 * │ before     │ after      │ mode           │
-	 * └────────────┴────────────┴────────────────┘
-	 *
-	 *  *before* – byte offset **after reservation** (writer fetch-add).
-	 *  *after*  – byte offset **after commit**     (writer release-store).
-	 *  *mode*   – current mode (read/follow/stop)
-	 */
-	struct alignas(64) RingMeta
-	{
-		std::atomic<uint64_t> before{};
-		std::atomic<uint64_t> after{};
-		std::atomic<RingMode> mode{RingMode::Stop};
-	};
-
 	/*
 	   DPDK ports used by `dataplane`.
 	   Each port has a name with which is exposed into host system
